Fix analogous int casting error in TieredHadoopTool
git-svn-id: https://gamescrafters.svn.sourceforge.net/svnroot/gamescrafters/GamesmanJava@4586 a668be5d-ef11-cb4c-8bfd-e08cbb24f2ba

+++ hadoop-src/edu/berkeley/gamesman/hadoop/TieredHadoopTool.java
@@ -95,8 +95,8 @@ public class TieredHadoopTool extends Configured implements Tool {
 			numMappers = 1;
 		}
 		if (minSplit > 0) {
-			if (((int)(endHash - firstHash))/numMappers < minSplit) {
-				numMappers = ((int)(endHash - firstHash))/minSplit;
+			if (((endHash - firstHash)/numMappers) < minSplit) {
+				numMappers = (int)((endHash - firstHash)/minSplit);
 				if (numMappers < 1) {
 					numMappers = 1;
 				}
